usecases: revise forecast trial examples
Revise the forecast trial section to add more specific, "real-world"
details such as it being an electric utility wanting to do the trial and
the focus being on forecasting power generation from the utility's fleet
of solar PV power plants.

Also add additional examples that build off of the first example. This
way readers can focus on the differences in the examples, rather than
having to re-read a copy&paste version of Example 1 and try to figure
out what the differences are.